Change buildah task to upload an SBOM blob
This change modifies the buildah task to upload the image SBOM as a blob
to an oci registry. It also emits a new result with a reference to the
SBOM blob, IMAGE_SBOM_URL.

NOTE: This is a work in progress. Ideally, the task can derive the
SBOM_REPO parameter from the IMAGE parameter by stripping off the
tag value if one is provided. Very doable. I just haven't gotten to it.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>