Use official tkn-cli image
This change modifies the tkn-bundle Task to use the officially supported
image containing the tkn cli.

The image is from registry.redhat.io which does require authentication.
This is fine when running the Task on OpenShift but it requires special
handling on other Kubernetes clusters, e.g. Kind.

The tkn-bundle Task is tested via shellspec. The tests execute the Task
on a local instance of a Kind cluster. For this reason, credentials are
plumbed through from the local host to the Kind cluster. NOTE: This only
affects testing, either locally or via GitHub workflows.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>